+++ server/src/main/java/org/opensearch/index/remote/RemoteStorePressureService.java
@@ -133,18 +133,18 @@ public class RemoteStorePressureService {
             if (pressureTracker.getRefreshSeqNoLag() <= 1) {
                 return true;
             }
-            if (pressureTracker.isUploadBytesAverageReady() == false) {
+            if (pressureTracker.isUploadBytesMovingAverageReady() == false) {
                 logger.trace("upload bytes moving average is not ready");
                 return true;
             }
-            double dynamicBytesLagThreshold = pressureTracker.getUploadBytesAverage() * pressureSettings.getBytesLagVarianceFactor();
+            double dynamicBytesLagThreshold = pressureTracker.getUploadBytesMovingAverage() * pressureSettings.getBytesLagVarianceFactor();
             long bytesLag = pressureTracker.getBytesLag();
             return bytesLag <= dynamicBytesLagThreshold;
         }
 
         @Override
         public String rejectionMessage(RemoteSegmentTransferTracker pressureTracker, ShardId shardId) {
-            double dynamicBytesLagThreshold = pressureTracker.getUploadBytesAverage() * pressureSettings.getBytesLagVarianceFactor();
+            double dynamicBytesLagThreshold = pressureTracker.getUploadBytesMovingAverage() * pressureSettings.getBytesLagVarianceFactor();
             return String.format(
                 Locale.ROOT,
                 "rejected execution on primary shard:%s due to remote segments lagging behind local segments."
@@ -179,18 +179,20 @@ public class RemoteStorePressureService {
             if (pressureTracker.getRefreshSeqNoLag() <= 1) {
                 return true;
             }
-            if (pressureTracker.isUploadTimeMsAverageReady() == false) {
+            if (pressureTracker.isUploadTimeMovingAverageReady() == false) {
                 logger.trace("upload time moving average is not ready");
                 return true;
             }
             long timeLag = pressureTracker.getTimeMsLag();
-            double dynamicTimeLagThreshold = pressureTracker.getUploadTimeMsAverage() * pressureSettings.getUploadTimeLagVarianceFactor();
+            double dynamicTimeLagThreshold = pressureTracker.getUploadTimeMovingAverage() * pressureSettings
+                .getUploadTimeLagVarianceFactor();
             return timeLag <= dynamicTimeLagThreshold;
         }
 
         @Override
         public String rejectionMessage(RemoteSegmentTransferTracker pressureTracker, ShardId shardId) {
-            double dynamicTimeLagThreshold = pressureTracker.getUploadTimeMsAverage() * pressureSettings.getUploadTimeLagVarianceFactor();
+            double dynamicTimeLagThreshold = pressureTracker.getUploadTimeMovingAverage() * pressureSettings
+                .getUploadTimeLagVarianceFactor();
             return String.format(
                 Locale.ROOT,
                 "rejected execution on primary shard:%s due to remote segments lagging behind local segments."